fix some warning

+++ src/ECS/ECS.cpp
@@ -35,7 +35,7 @@ Entity Registry::CreateEntity() {
 	entity.registry = this;
 	entitiesToBeAdded.insert(entity);
 
-	if (entityId >= entityComponentSignatures.size()) {
+	if (entityId >= static_cast<int>(entityComponentSignatures.size())) {
 		entityComponentSignatures.resize(entityId + 1);
 	}
